pyproject.toml
[project]
name = "mcp-mixsearch"
version = "0.1.0"
description = "Core infrastructure for building MCP servers with web search capabilities"
readme = "README.md"
requires-python = ">=3.13"
license = {file = "LICENSE"}
authors = [
    {name = "Geovanny F. Fajardo", email = "gffajardo@gmail.com"}
]
keywords = [
    "mcp",
    "model-context-protocol",
    "web-search",
    "fastmcp",
    "ai-tools"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    "fastapi>=0.104.0",
    "fastmcp>=2.0.0",
    "pydantic>=2.0.0",
    "playwright>=1.48.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.1.0",
    "httpx>=0.25.0",
    "fake-useragent>=1.4.0",
    "mcp-weather @ git+https://github.com/geosp/mcp-weather.git",
    "ddgs>=9.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
mcp-mixsearch = "server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
[tool.hatch.metadata]
allow-direct-references = true
